% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\value{
The sum of x and y
}
\description{
Add two numbers
}
\examples{
add(1, 2)
add(2, 3)
}
